requirements.txt
aiohttp
annotated-types==0.7.0
anyio==4.8.0
beanie
boto3
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.5
chromadb==0.5.4
distro==1.9.0
faiss-cpu
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
langchain==0.2.17
langchain-chroma==0.1.2
langchain-core==0.2.43
langchain-openai==0.1.25
langfuse==2.57.2
motor
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.61.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pymongo==4.9.2
python-decouple==3.8
python-dotenv==1.0.1
python-multipart
requests==2.32.3
sniffio==1.3.1
tenacity==8.5.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==11.0.3
Pillow
cryptography
cloudinary